requirements.txt
moviepy==1.0.3
TTS==0.16.6
mecab-python3==1.0.6
unidic-lite==1.0.8
python-mpd2==3.1.0
nicegui==1.3.13
honcho==1.1.0
